% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ars.R
\name{add_cord_lines}
\alias{add_cord_lines}
\title{compute_h_u}
\usage{
add_cord_lines(
  norm_Y_bar = 0.8,
  n = 10,
  nu = 1,
  X_val = NULL,
  left_point = NULL,
  right_point = NULL,
  inflex = NULL,
  left_most_x = 1e-16
)
}
\description{
compute_h_u
}
\examples{
compute_h_u(orm_Y_bar=.8, n=10, nu=1)
}
