% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{Finite_mixture_of_VONF}
\alias{Finite_mixture_of_VONF}
\title{Finite_mixture_of_VONF}
\usage{
Finite_mixture_of_VONF(Y, K, max_iter = 100, starting_val = NULL)
}
\description{
Finite_mixture_of_VONF
}
\examples{
data=Generate_mix_of_VMF(n=1000, K = 4)
Y=data$data
mix_vonf_fit<-Finite_mixture_of_VONF(Y, K=4)
}
