% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{Generate_mix_of_VMF}
\alias{Generate_mix_of_VMF}
\title{Generate_mix_of_VMF}
\usage{
Generate_mix_of_VMF(
  n,
  K,
  mean_direction_list = NULL,
  concentration_list = NULL
)
}
\description{
Generate_mix_of_VMF
}
\examples{
data_par=Generate_mix_of_VMF(100, K=1)
}
