% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{Generate_random_direction}
\alias{Generate_random_direction}
\title{export
Generate_random_direction}
\usage{
Generate_random_direction(space_dim)
}
\description{
export
Generate_random_direction
}
\examples{
#' Norm_vec(Generate_random_direction(3)) #checking the accuracy
}
