% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{initial_cluster_generation}
\alias{initial_cluster_generation}
\title{initial_cluster_generation}
\usage{
initial_cluster_generation(Y, cluster_size)
}
\description{
initial_cluster_generation
}
