% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_Approximations.R
\name{log_Bessel_K}
\alias{log_Bessel_K}
\title{calculates log(K(x, nu)) where K is the modified Bessel function of the Second kind.}
\usage{
log_Bessel_K(x, nu)
}
\arguments{
\item{x}{A complex number.}

\item{nu}{degree of the Bessel function}
}
\value{
The log(K(\code{x},\code{nu}))
}
\description{
calculates log(K(x, nu)) where K is the modified Bessel function of the Second kind.
}
\examples{
log_Bessel_K(10,.5)
log_Bessel_K(10+3*1i,0)
 log_Bessel_K(10000,0)
}
