% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_Approximations.R
\name{log_sinh}
\alias{log_sinh}
\title{calculates log(sinh(x)) for small and large arguments.}
\usage{
log_sinh(x)
}
\arguments{
\item{x}{A real number, a vector of real numbers.}
}
\value{
The log(sinh(\code{x}))
}
\description{
calculates log(sinh(x)) for small and large arguments.
}
\examples{
log_sinh(10)
log_sinh(c(1000,1,10,10000))
}
