% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_Approximations.R
\name{plot_g_function}
\alias{plot_g_function}
\title{plot_marginl_post}
\usage{
plot_g_function(fun, par, x_lim = c(1e-07, 5), par_list)
}
\description{
plot_marginl_post
}
\examples{
par_list=list(nu=1,norm_Y_bar=.9, n=10,ifLog=TRUE)
plot_marginl_post(nu=.5,norm_Y_bar = .8, x_lim = c(0,10), ifLog = TRUE)
plot_g_function(fun=Marginal_post_kappa,par_list=par_list )
par_list1=list(nu=1,norm_Y_bar=.9, n=10)
plot_g_function(fun=Marginal_log_post_kappa_prime_prime,par_list=par_list1 )

}
