% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ars.R
\name{plot_h_and_h_u}
\alias{plot_h_and_h_u}
\title{plot_h_and_h_u}
\usage{
plot_h_and_h_u(norm_Y_bar = 0.8, n = 10, nu = 1, Ulim = 20)
}
\description{
plot_h_and_h_u
}
\examples{
plot_h_and_h_u(norm_Y_bar=.8, n=10, nu=1,Ulim=20)
}
