% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_Approximations.R
\name{plot_marginl_post}
\alias{plot_marginl_post}
\title{plot_marginl_post}
\usage{
plot_marginl_post(
  nu = 0,
  norm_Y_bar = 0.9,
  n = 10,
  x_lim = c(1e-07, 5),
  ifLog = FALSE
)
}
\description{
plot_marginl_post
}
\examples{
plot_marginl_post(nu=.5,norm_Y_bar = .3, x_lim = c(.00001,10), ifLog = TRUE)
VCLUST::plot_marginl_post(nu=.5,norm_Y_bar = .8, x_lim = c(0,10), ifLog = TRUE)
}
