% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound.r
\name{rGenSample}
\alias{rGenSample}
\title{rGenSample(N=10, norm_Y_bar = .7,nu = 1,n = 100)}
\usage{
rGenSample(
  N = 10,
  norm_Y_bar = 0.7,
  nu = 1,
  n = 100,
  ifPlot = TRUE,
  initial_num_of_points_get_inflex = 4,
  method = "UpdateALL",
  BugTest1 = TRUE,
  Ulim = 20,
  sleep_time = 4,
  newPlot = FALSE,
  Version_Init_Gen = 1
)
}
\arguments{
\item{N}{: Number of sample to be Generated}

\item{norm_Y_bar}{: Norm of the Response Directional Vector. Note that norm_Y_bar is a number between 0 and 1.}

\item{nu}{: nu=d/2-1 where the directional responses belongs to S^{d-1}. For circular data d=2 and \nu= \frac{1}{2}}

\item{n}{: Sample size for the collected data.}

\item{ifPlot=TRUE}{: If we want to plot/visualize the log marginal densityand the log of the proposal bounds.}

\item{newPlot=FALSE}{: There are two types of plot. newPlot=TRUE provides more fancy plot but requires more time and the package ggplot2}

\item{initial_num_of_points_get_inflex=4}{: The role of initial_num_of_points_ greater than the inflex point  is number of points to start creating the proposal. We may increase efficiency by choosing appropriate number. If we want to generate a large number of samples for example,  10000, we may choose this parameter to be 50 or 100 for a more time efficient sampling.}

\item{sleep_time=4}{: When ifPlot=TRUE, the sleep_time is the pause amount to display the adaptive change in the corresponding display.}
}
\description{
rGenSample(N=10, norm_Y_bar = .7,nu = 1,n = 100)
}
\examples{
rGenSample(N=10, norm_Y_bar = .7,nu = 1,n = 100, ifPlot = FALSE, BugTest1 = FALSE)
rGenSample(N=10, norm_Y_bar = .7,nu = 1,n = 100, ifPlot = TRUE,newPlot=FALSE , BugTest1 = FALSE, Ulim=15, sleep_time=4)
rGenSample(N=10, norm_Y_bar = .7,nu = 1,n = 100, ifPlot = TRUE,newPlot=TRUE , BugTest1 = FALSE, Ulim=15, sleep_time=4)

}
\author{
Dr. Subhadip Pal, UAE University
}
