% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{Update_parameter_for_single_cluster}
\alias{Update_parameter_for_single_cluster}
\title{Update_parameter_for_single_cluster}
\usage{
Update_parameter_for_single_cluster(sel_data, curr_param_j_th, hyper = NULL)
}
\description{
Update_parameter_for_single_cluster
}
\examples{
data_var=Generate_mix_of_VMF(n=100, K = 1)
data=data_var$data
param=list(mu=unlist(data_var$True_parameter$mean_direction_list),
           kappa=unlist(data_var$True_parameter$concentration_list))
Update_parameter_for_single_cluster(sel_data=data,curr_param_j_th=param )
}
