% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ars.R
\name{Update_tangent_lines_recalculate_all}
\alias{Update_tangent_lines_recalculate_all}
\title{Update_tangent_lines_recalculate_all  returns the same output as Update_tangent_lines
Update_tangent_lines_recalculate_all calculates the lines all over again
whereas Update_tangent_lines adds the new one keeping the old lines}
\usage{
Update_tangent_lines_recalculate_all(
  norm_Y_bar = 0.8,
  n = 10,
  nu = 1,
  old_lines = NULL,
  xnew = xnew
)
}
\description{
Update_tangent_lines_recalculate_all  returns the same output as Update_tangent_lines
Update_tangent_lines_recalculate_all calculates the lines all over again
whereas Update_tangent_lines adds the new one keeping the old lines
}
