% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_VONF_functions_V2.R
\name{updateZ}
\alias{updateZ}
\title{Updating Z (Cluster membership) posterior#####}
\usage{
updateZ(Y, mu_all_cluster, kappa_all_cluster, pi_vec)
}
\description{
Updating Z (Cluster membership) posterior#####
}
