% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validator}
\alias{validator}
\title{Construct a Validator}
\usage{
validator(
  validator_cutpoint = NULL,
  validator_cutpoint_sign = NULL,
  validator_cutpoint_max_min_mean = NULL,
  predicted_cluster_combination = NULL,
  predicted_cluster_n = NULL,
  validator_source_variables,
  listwise_deletion_variables = NULL,
  validator_type,
  validator_position = NULL,
  flip_outcome_type = NULL,
  flipped_predictors_variables = NULL,
  flipped_predictors_cluster = NULL,
  flipped_predictors_pp = NULL,
  supervised_model = "logistic",
  logistic = TRUE,
  alpha = 1,
  lambda = 0,
  seed_num = c(seed_num_split = NA, seed_num_kfold = NA, seed_num_supervised_model =
    NA),
  validator_source_all_missing = 0
)
}
\value{
return a validator object
}
\description{
Construct a Validator
}
